Ion Matrix Calculation

Calculates conductivity contribution from provided ion concentrations. Uses Kohlrausch's Law. If a total conductivity measurement is not provided the total conductivity is calculated as the sum of all ions present in the data. If the user provides a conductivity measurement then ion contributions are a percentage of that number. In this case "Other" is added as an ion to capture any percentage of total conductivity not represented by the provided ions. Plotting is done outside of this function.

IonContrib(ion.data, ion.ref = MBSS.Ion.Ref, col.Cond = "")

Arguments

ion.data
Input data. Column names match rows in the ref table

ion.ref
Reference data for ions; Name (matches columns in data), Multiplier, and SortOrder. Default is included ref table.

col.Cond
Column name with conductivity in "data". Default is blank. If blank only ions present are used to calculate total conductivity for samples.

Value

Returns a data frame with additional fields for "multiplier", "ion conductivity", "total conductivity", "percent contribution" (per ion).

Details

No special libraries are needed for this function.

Examples