Taxa Observation Maps

Map taxonomic observations in Maryland. Inputs are 3 GIS shapefiles (State, County, and Hydrology) for Maryland, Observations (taxa names and Lat/Longs), and #' a crosswalk table of taxa names and map names. Input files are assumed to be Excel with the data on the first worksheet.

MapTaxaObs(
  obs,
  xWalk,
  dirMain = getwd(),
  onlymatches = TRUE,
  dirData = "Data",
  dirGIS = "GIS",
  dirMaps = "Maps",
  verbose = TRUE
)

Arguments

obs
Excel file containing 3 columns: TaxaName, Latitude83, Longitude83

xWalk
Crosswalk (Excel file) of taxa names between taxa name in occurence file and file name outputs for maps. Three columns (CommonName, Species, MapName).

dirMain
Path for folder containing obs and xwalk. Optional. Default = getwd().

onlymatches
Flag for mapping obs taxa matching xwalk taxa, or all taxa. Optional. Default = TRUE.

dirData
Subdirectory of "dir" specified above with data files. Default = "Data"

dirGIS
Subdirectory of "dir" specified above with GIS files. Default = "GIS"

dirMaps
Subdirectory of "dir" specified above for map outputs. Default = "Maps"

verbose
Boolean value for including feedback to the user that the script is working. Default = TRUE.

Value

Saves the maps and a dataframe of observed taxa not matching xwalk taxa to the dirMaps folder.

Details

The example data is fish but can be used for benthic macroinvertebrates as well. Maps are output as PNG. There is a flag for "only" matches between the data and the crosswalk table (this is the default). There is also a "verbose" flag for printing status messages to the user (this is the default). Tests for existence of Data, GIS directories and required files. On failure, stops and prints missing directories/files to screen.

The R libraries rgdal and readxl are required for this function.

Examples

# Set Working Directory
wd <- getwd()
# Create Example Data if Needed
## Create Directories
myDir.create <- file.path(wd,"Data")
ifelse(dir.exists(myDir.create)==FALSE,dir.create(myDir.create),"Directory
already exists")

#> [1] "Directory\nalready exists"

myDir.create <- file.path(wd,"GIS")
ifelse(dir.exists(myDir.create)==FALSE,dir.create(myDir.create),"Directory
already exists")

#> [1] "Directory\nalready exists"

myDir.create <- file.path(wd,"Maps")
ifelse(dir.exists(myDir.create)==FALSE,dir.create(myDir.create),"Directory
already exists")

#> [1] TRUE

## Create Data
### Taxa Data
myFiles <- c("AllFish_95to16.xls", "TaxaMapsCrossWalk20170731.xlsx")
file.copy(file.path(path.package("MBSStools"),"extdata",myFiles),
file.path(wd,"Data",myFiles))

#> [1] FALSE FALSE

### GIS
unzip(file.path(path.package("MBSStools"),"extdata","MD_GIS.zip"),
exdir=file.path(wd,"GIS"))

# Inputs
Obs <- "AllFish_95to16.xls"
XWalk <- "TaxaMapsCrossWalk20170731.xlsx"
wd <- getwd()

# Create maps
MapTaxaObs(Obs, XWalk, wd)

#> /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpEhbyva/file7741716e0ec/reference

#> OGR data source with driver: ESRI Shapefile 
#> Source: "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpEhbyva/file7741716e0ec/reference/GIS", layer: "MD_State_Boundary"
#> with 2 features
#> It has 9 fields
#> Integer64 fields read as strings:  OBJECTID 
#> OGR data source with driver: ESRI Shapefile 
#> Source: "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpEhbyva/file7741716e0ec/reference/GIS", layer: "MD_Coast_Hydrology"
#> with 1117 features
#> It has 9 fields
#> Integer64 fields read as strings:  OBJECTID 
#> OGR data source with driver: ESRI Shapefile 
#> Source: "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpEhbyva/file7741716e0ec/reference/GIS", layer: "MD_Boundary_County_Detailed"
#> with 24 features
#> It has 7 fields
#> Integer64 fields read as strings:  OBJECTID DISTRICT COUNTY_FIP COUNTYNUM 

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'